src/requirements.txt
email-validator==2.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Jinja2==3.1.6
python-dotenv==1.1.1
psycopg2==2.9.11
WTForms==3.2.1